refactor: create a minimal topnav comopnent to be used by error views
