+++ src/components/AppComponents/ErrorBoundary/ErrorBoundary.tsx
@@ -8,7 +8,7 @@ import getBrowserConfig from "@/browser-lib/getBrowserConfig";
 import { defaultTopNavProps } from "@/pages/_error";
 
 const ClientErrorView: FC = () => {
-  return <ErrorView footerVariant="client-error" topNav={defaultTopNavProps} />;
+  return <ErrorView topNav={defaultTopNavProps} />;
 };
 
 class NonBugsnagErrorBoundary extends Component<
